updated data imports to database (planets/trip)

+++ scripts/import_planets.php
@@ -1,49 +1,55 @@
 <?php
 
 use Travia\Classes\Planet;
-require_once "../class/Planet.php";
+use Travia\Classes\Trip;
 
-function import_planets(string $json)
+require_once "../class/Planet.php";
+require_once "../class/Trip.php";
+
+function import_planets(string $jsonFilePath)
 {
     // Connection to database
     global $dbh;
     include("../utils/connection.php");
 
-    echo "here";
-
     // Import data from json file
-    //$jsonData = json_decode(file_get_contents($json), true);
-
-    $jsonContent = file_get_contents($json);
-
-    echo "<br>jsonContent ";
-
-    // SOMEHOW NOT WORKING HERE
+    ini_set('memory_limit', '512M'); // Increase memory limit
+    $jsonContent = file_get_contents($jsonFilePath);
     $jsonData = json_decode($jsonContent, true);
-
-    echo "<br>jsonData<br>";
+    if ($jsonData === null) {
+        echo "Error : invalid json file";
+        return;
+    }
 
     try {
-        // Begin transaction
-        $dbh->beginTransaction();
 
-        // Empty table before filling
+        // Empty tables before filling
         $dbh->exec("TRUNCATE TABLE TRAVIA_Planet;");
+        $dbh->exec("TRUNCATE TABLE TRAVIA_Trip;");
 
         // Prepare statement for insert
         $stmt = $dbh->prepare(
             "INSERT INTO TRAVIA_Planet (id, name, image, coord, x, y, subGridCoord, subGridX, subGridY, sunName, region, sector, suns, moons, position, distance, dayLength, yearLength, diameter, gravity)
              VALUES (:id, :name, :image, :coord, :x, :y, :subGridCoord, :subGridX, :subGridY, :sunName, :region, :sector, :suns, :moons, :position, :distance, :dayLength, :yearLength, :diameter, :gravity);"
         );
+        $stmt2 = $dbh->prepare(
+            "INSERT INTO TRAVIA_Ship (departurePlanet, destinationPlanet, day, departureTime, ship)
+            VALUES (:departurePlanet, :destinationPlanet, :day, :time, :ship);"
+        );
 
         $count = 0;
 
-        echo "one";
         // Iterate over JSON file
         foreach ($jsonData as $planetData) {
 
-            // If planet is successfully created, valid data
-            $planet = new Planet(
+            // PLANET : Set missing values
+            $planetData["SunName"] ??= "";
+            $planetData["Image"] ??= "";
+            $planetData["SubGridCoord"] ??= "";
+            $planetData["Coord"] ??= "";
+
+            // PLANET : If planet is successfully created, data is valid
+            $planetObject = new Planet(
                 $planetData["Id"],
                 $planetData["Name"],
                 $planetData["Image"],
@@ -66,51 +72,73 @@ function import_planets(string $json)
                 $planetData["Gravity"]
             );
 
-            // Counter
-            $count++;
-            if ( $count%50 == 0 ) {
-                echo "<br>";
-            }
-            echo $count . " ";
-
-            // Set values if null
-            $image = $planetData["Image"] ?? "";
-
-            // Bind parameters and execute (data)
-            $stmt->bindParam(":id", $planetData["Id"], PDO::PARAM_INT);
-            $stmt->bindParam(":name", $planetData["Name"], PDO::PARAM_STR);
-            $stmt->bindParam(":image", $image, PDO::PARAM_STR);
-            $stmt->bindParam(":coord", $planetData["Coord"], PDO::PARAM_STR);
-            $stmt->bindParam(":x", $planetData["X"], PDO::PARAM_INT);
-            $stmt->bindParam(":y", $planetData["Y"], PDO::PARAM_INT);
-            $stmt->bindParam(":subGridCoord", $planetData["SubGridCoord"], PDO::PARAM_STR);
-            $stmt->bindParam(":subGridX", $planetData["SubGridX"], PDO::PARAM_STR);  // processed as float
-            $stmt->bindParam(":subGridY", $planetData["SubGridY"], PDO::PARAM_STR); // processed as float
-            $stmt->bindParam(":sunName", $planetData["SunName"], PDO::PARAM_STR);
-            $stmt->bindParam(":region", $planetData["Region"], PDO::PARAM_STR);
-            $stmt->bindParam(":sector", $planetData["Sector"], PDO::PARAM_STR);
-            $stmt->bindParam(":suns", $planetData["Suns"], PDO::PARAM_INT);
-            $stmt->bindParam(":moons", $planetData["Moons"], PDO::PARAM_INT);
-            $stmt->bindParam(":position", $planetData["Position"], PDO::PARAM_INT);
-            $stmt->bindParam(":distance", $planetData["Distance"], PDO::PARAM_INT);
-            $stmt->bindParam(":dayLength", $planetData["LengthDay"], PDO::PARAM_INT);
-            $stmt->bindParam(":yearLength", $planetData["LengthYear"], PDO::PARAM_INT);
-            $stmt->bindParam(":diameter", $planetData["Diameter"], PDO::PARAM_INT);
-            $stmt->bindParam(":gravity", $planetData["Gravity"], PDO::PARAM_INT);
+            // PLANET : Bind values + add planet to database
+            $stmt->bindValue(":id", $planetData["Id"], PDO::PARAM_INT);
+            $stmt->bindValue(":name", $planetData["Name"], PDO::PARAM_STR);
+            $stmt->bindValue(":image", $planetData["Image"], PDO::PARAM_STR);
+            $stmt->bindValue(":coord", $planetData["Coord"], PDO::PARAM_STR);
+            $stmt->bindValue(":x", $planetData["X"], PDO::PARAM_INT);
+            $stmt->bindValue(":y", $planetData["Y"], PDO::PARAM_INT);
+            $stmt->bindValue(":subGridCoord", $planetData["SubGridCoord"], PDO::PARAM_STR);
+            $stmt->bindValue(":subGridX", $planetData["SubGridX"], PDO::PARAM_STR);  // processed as float
+            $stmt->bindValue(":subGridY", $planetData["SubGridY"], PDO::PARAM_STR); // processed as float
+            $stmt->bindValue(":sunName", $planetData["SunName"], PDO::PARAM_STR);
+            $stmt->bindValue(":region", $planetData["Region"], PDO::PARAM_STR);
+            $stmt->bindValue(":sector", $planetData["Sector"], PDO::PARAM_STR);
+            $stmt->bindValue(":suns", $planetData["Suns"], PDO::PARAM_INT);
+            $stmt->bindValue(":moons", $planetData["Moons"], PDO::PARAM_INT);
+            $stmt->bindValue(":position", $planetData["Position"], PDO::PARAM_INT);
+            $stmt->bindValue(":distance", $planetData["Distance"], PDO::PARAM_INT);
+            $stmt->bindValue(":dayLength", $planetData["LengthDay"], PDO::PARAM_INT);
+            $stmt->bindValue(":yearLength", $planetData["LengthYear"], PDO::PARAM_INT);
+            $stmt->bindValue(":diameter", $planetData["Diameter"], PDO::PARAM_INT);
+            $stmt->bindValue(":gravity", $planetData["Gravity"], PDO::PARAM_INT);
             $stmt->execute();
 
+
+            // TRIPS
+            foreach ($planetData["trips"] as $day => $trip) {
+
+                foreach ($trip as $tripData) {
+
+                    // If trip is successfully set, data is valid
+                    $tripObject = new Trip(
+                        $planetData["Id"],
+                        $tripData["destination_planet_id"][0],
+                        $day,
+                        $tripData["departure_time"][0],
+                        $tripData["ship_id"][0]
+                    );
+
+                    // TRIP : Bind values + add trip to database
+                    $stmt2->bindValue(":departurePlanet", $planetData["Id"], PDO::PARAM_INT);
+                    $stmt2->bindValue(":destinationPlanet", $tripData["destination_planet_id"][0], PDO::PARAM_INT);
+                    $stmt2->bindValue(":day", $day, PDO::PARAM_STR);
+                    $stmt2->bindValue(":time", $tripData["time"][0], PDO::PARAM_STR);
+                    $stmt2->bindValue(":ship", $tripData["ship_id"][0], PDO::PARAM_INT);
+                    $stmt2->execute();
+                }
+            }
+
             // Set unused object reference to null for garbage collection
-            $planet = null;
+            $planetObject = null;
+            $tripObject = null;
+
+            // Counter
+            if ($count%50 == 0) { // Jump a line every 50 entries
+                echo "<br>";
+            }
+            echo ++$count . " ";
+
         }
 
-        // Commit transaction + trigger garbage collection
-        $dbh->commit();
         echo "<br>Planets successfully updated.<br>";
+
+        // Trigger garbage collection
         gc_collect_cycles();
 
 
     } catch (Exception $e) {
-        $dbh->rollBack();
         echo "Error while importing planets to database : " . $e->getMessage();
     }
 }
